requirements.txt
modal
requests
Pillow
fastapi
pydantic